Run type system verification tests for non-canary mode as well
